+++ lifelines/compat.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-import sys
-
-PY2 = sys.version_info[0] == 2
-PY3 = sys.version_info[0] >= 3