+++ lifelines/fitters/cox_time_varying_fitter.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
-from __future__ import print_function
-from __future__ import division
+
 
 from datetime import datetime
 import warnings
@@ -82,8 +81,7 @@ class CoxTimeVaryingFitter(BaseFitter):
     """
 
     def __init__(self, alpha=0.05, penalizer=0.0, strata=None):
-        if not (0 < alpha <= 1.0):
-            raise ValueError("alpha parameter must be between 0 and 1.")
+        super(CoxTimeVaryingFitter, self).__init__(alpha=alpha)
         if penalizer < 0:
             raise ValueError("penalizer parameter must be >= 0.")
 
@@ -103,6 +101,7 @@ class CoxTimeVaryingFitter(BaseFitter):
         step_size=None,
         robust=False,
         strata=None,
+        initial_point=None,
     ):  # pylint: disable=too-many-arguments
         """
         Fit the Cox Proportional Hazard model to a time varying dataset. Tied survival times
@@ -135,7 +134,14 @@ class CoxTimeVaryingFitter(BaseFitter):
           "The Robust Inference for the Cox Proportional Hazards Model", Journal of the American Statistical Association, Vol. 84, No. 408 (Dec., 1989), pp. 1074- 1078
         step_size: float, optional
             set an initial step size for the fitting algorithm.
-        strata: TODO
+        strata: list or string, optional
+            specify a column or list of columns n to use in stratification. This is useful if a
+            categorical covariate does not obey the proportional hazard assumption. This
+            is used similar to the `strata` expression in R.
+            See http://courses.washington.edu/b515/l17.pdf.
+        initial_point: (d,) numpy array, optional
+            initialize the starting point of the iterative
+            algorithm. Default is the zero vector.
 
         Returns
         --------
@@ -154,7 +160,7 @@ class CoxTimeVaryingFitter(BaseFitter):
         df = df.copy()
 
         if not (id_col in df and event_col in df and start_col in df and stop_col in df):
-            raise KeyError("A column specified in the call to `fit` does not exist in the dataframe provided.")
+            raise KeyError("A column specified in the call to `fit` does not exist in the DataFrame provided.")
 
         if weights_col is None:
             self.weights_col = None
@@ -196,6 +202,7 @@ class CoxTimeVaryingFitter(BaseFitter):
             start,
             stop,
             weights,
+            initial_point=initial_point,
             show_progress=show_progress,
             step_size=step_size,
         )
@@ -285,7 +292,17 @@ class CoxTimeVaryingFitter(BaseFitter):
             return df
 
     def _newton_rhaphson(
-        self, df, events, start, stop, weights, show_progress=False, step_size=None, precision=10e-6, max_steps=50
+        self,
+        df,
+        events,
+        start,
+        stop,
+        weights,
+        show_progress=False,
+        step_size=None,
+        precision=10e-6,
+        max_steps=50,
+        initial_point=None,
     ):  # pylint: disable=too-many-arguments,too-many-locals,too-many-branches,too-many-statements
         """
         Newton Rhaphson algorithm for fitting CPH model.
@@ -312,7 +329,10 @@ class CoxTimeVaryingFitter(BaseFitter):
         _, d = df.shape
 
         # make sure betas are correct size.
-        beta = np.zeros((d,))
+        if initial_point is not None:
+            beta = initial_point
+        else:
+            beta = np.zeros((d,))
 
         i = 0
         converging = True
@@ -403,7 +423,7 @@ https://lifelines.readthedocs.io/en/latest/Examples.html#problems-with-convergen
             elif abs(ll) < 0.0001 and norm_delta > 1.0:
                 warnings.warn(
                     "The log-likelihood is getting suspiciously close to 0 and the delta is still large. There may be complete separation in the dataset. This may result in incorrect inference of coefficients. \
-See https://stats.idre.ucla.edu/other/mult-pkg/faq/general/faqwhat-is-complete-or-quasi-complete-separation-in-logisticprobit-regression-and-how-do-we-deal-with-them/ ",
+See https://stats.stackexchange.com/questions/11109/how-to-deal-with-perfect-separation-in-logistic-regression",
                     ConvergenceWarning,
                 )
                 converging, completed = False, False