+++ lifelines/fitters/log_logistic_aft_fitter.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from __future__ import division
+
 
 from autograd import numpy as np
 import pandas as pd