+++ lifelines/fitters/nelson_aalen_fitter.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
-from __future__ import print_function
-from __future__ import division
+
 
 import warnings
 import numpy as np