+++ tests/test_generate_datasets.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-# test_generate_datasets.py
-from __future__ import print_function
 import os
 
 import pytest