+++ tests/test_statistics.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import print_function
 import numpy as np
 import pandas as pd
 import numpy.testing as npt